save update migrate CustomerTable

+++ Modules/Customer/Database/Migrations/2024_10_13_082720_create_customers_table.php
@@ -11,23 +11,27 @@ return new class extends Migration
      */
     public function up(): void
     {
-        Schema::create('customers', function (Blueprint $table) {
-            $table->id();
-            $table->unsignedBigInteger('group_id')->comment('Mã Nhóm Khách hàng');
-            $table->string('name')->nullable()->comment('Tên khách hàng');
-            $table->integer('phone')->nullable()->comment('Số điện thoại');
-            $table->string('email')->unique()->comment('Email');
-            $table->string('address')->nullable()->comment('Địa chỉ');
-            $table->integer('point')->nullable()->default(0)->comment('Điểm thành viên');
-            $table->tinyInteger('status')->comment('Trạng thái tài khoản');
-            $table->string('password')->comment('Mật khẩu');
-            $table->timestamp('email_verified_at')->nullable();
-            $table->rememberToken();
-            $table->softDeletes();
-            $table->timestamps();
+        if (!Schema::hasTable('customers')) {
+            Schema::create('customers', function (Blueprint $table) {
+                $table->id();
+                $table->unsignedBigInteger('group_id')->comment('Mã Nhóm Khách hàng');
+                $table->string('name')->nullable()->comment('Tên khách hàng');
+                $table->integer('phone')->nullable()->comment('Số điện thoại');
+                $table->string('email')->unique()->comment('Email');
+                $table->string('address')->nullable()->comment('Địa chỉ');
+                $table->integer('point')->nullable()->default(0)->comment('Điểm thành viên');
+                $table->tinyInteger('status')->comment('Trạng thái tài khoản');
+                $table->string('password')->comment('Mật khẩu');
+                $table->timestamp('email_verified_at')->nullable();
+                $table->rememberToken();
+                $table->softDeletes();
+                $table->timestamps();
 
-            $table->foreign('group_id')->references('id')->on('group_customer')->onDelete('cascade');
-        });
+                $table->foreign('group_id')->references('id')->on('group_customer')->onDelete('cascade');
+            });
+        } else {
+            Schema::table('customers', function (Blueprint $table) {});
+        }
     }
 
     /**
